+++ examples/sw_attention_lazy.py
@@ -105,7 +105,8 @@ class SWAttention(nn.Module):
         self.bias = nn.Parameter(torch.zeros(self.num_heads, self.max_bias_length))
 
         # 每个 head 的稀疏性参数: [num_heads]
-        self.tau = nn.Parameter(torch.full((self.num_heads,), -1.0))
+        # Use float32 to avoid gradient underflow with bfloat16 training
+        self.tau = nn.Parameter(torch.full((self.num_heads,), -1.0, dtype=torch.float32))
 
         self.rotary = RotaryEmbedding(dim=self.head_dim, base=self.rope_theta)
 
@@ -172,10 +173,11 @@ class SWAttention(nn.Module):
                 varlen = attention_mask.sum(dim=1).to(torch.int32)
 
         # ========== 关键修改 7: 调用 Lazy Attention ==========
+        # Convert tau from float32 to model dtype for kernel computation
         attn_output = lazy_attention_triton(
             q, k, v,
             bias=self.bias,
-            tau=self.tau,
+            tau=self.tau.to(q.dtype),
             window_size=self.max_bias_length,
             varlen=varlen
         )
